+++ isa-project/src/main/java/isaproject/model/boat/Boat.java
@@ -61,7 +61,7 @@ public class Boat implements Serializable {
 	@ElementCollection(fetch = FetchType.EAGER)
 	@CollectionTable(name = "boat_fishing_equipment", joinColumns = @JoinColumn(name = "boat_id"), foreignKey = @ForeignKey(name = "equipment_boat"))
 	private Set<String> fishingEquipment = new HashSet<>();
-	private Integer pricePerHour;
+	private Double pricePerHour;
 	@JsonManagedReference
 	@OneToMany(mappedBy = "boat", fetch = FetchType.EAGER)
 	private Set<BoatQuickReservation> boatQuickReservation = new HashSet<>();
